requirements/requirements.txt
electrumsv-node>=0.0.21
requests
filelock>=3.3.2
psutil-wheels>=5.6.6
setuptools>=49.6.0
aiorpcX>=0.22.1
bitcoinX>=0.2.4
peewee>=3.13.3
wheel
aiohttp>=3.7.3
tailer
stringcase
colorama
python-dotenv
asyncpg
pylru
pytest
aiodns
